dataserve/requirements.txt
-i https://pypi.org/simple
boto3==1.9.106
botocore==1.12.106
certifi==2018.11.29
chardet==3.0.4
click==7.0
clickclick==1.2.2
connexion==2018.0.dev1
docutils==0.14
flask==1.0.2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
markupsafe==1.1.1
openapi-spec-validator==0.2.6
paho-mqtt==1.4.0
pathlib==1.0.1
python-dateutil==2.8.0 ; python_version >= '2.7'
pyyaml==5.1b3
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1 ; python_version >= '3.4'
werkzeug==0.14.1
